add login & permission

+++ antplatform-admin-biz/src/main/java/com/antplatform/admin/biz/infrastructure/tkmybatis/ext/BaseMapper.java
@@ -12,7 +12,7 @@ import tk.mybatis.mapper.common.special.InsertListMapper;
 public interface BaseMapper<T> extends
         Mapper<T>,
         InsertListMapper<T>,
-        ExpandedIdListMapper<T, Long>,
-        UpdateIdListMapper<T, Long>,
+        ExpandedIdListMapper<T, Integer>,
+        UpdateIdListMapper<T, Integer>,
         PageModelMapper<T> {
 }
